refactor(core): consolidate current-user schemas into schemata/types
currentUserTypes.ts held zod schemas under a "types" name -- the one
file that broke the schemata-in-users.schemata / types-in-users.types
split the rest of the module follows. Move the signed-in/out and
discriminated-union schemas into users.schemata.ts beside their
siblings, the inferred types into users.types.ts, and delete
currentUserTypes.ts plus the redundant ./types barrel. Public exports
are unchanged.

Also discriminate DonorTeacherBannerDetailsSchema on user_type (the
transform moves to the union, since zod can't discriminate through a
ZodEffects) so a malformed payload fails on the discriminant rather
than as an opaque union mismatch.
