+++ frontend/packages/core/src/api/dashboard/users/index.ts
@@ -3,4 +3,3 @@ export * from './users.keys';
 export * from './users.query';
 export * from './users.schemata';
 export * from './users.types';
-export * from './types';
