--- frontend/packages/core/src/api/dashboard/users/types.ts
@@ -1,1 +1,0 @@
-export * from './currentUserTypes';
